sentilytics/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.0
textblob==0.17.1
vaderSentiment==3.3.2
nltk==3.8.1
python-dotenv==1.0.0
aiohttp==3.9.1
